Overview map and TDB generation.
We are going to allow several files to be supplied on the command line and build up the tdb and the
overview map as we go along.

In the future we will be able to read img files and create the files that way.

A special map data source is being used to collect the overview information.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@277 25d90789-57f7-4ee0-8453-03a3dfeeeb22