add OffsetWriter to RoadDef, and RaodDef to Polyline
RoadDefs now have an OffsetWriterList: If a RoadDef knows
its final offset, it calls the owlist, so that all the
pointers get updated.

Polylines now have a link to a RoadDef.
When the Polyline gets written out (which happens quiet
early), it puts the Label offset as a pointer for starters,
so that something valid is in there. Then it registers an
OffsetWriter in the RoadDef, so that the pointer gets
updated later on.


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@606 25d90789-57f7-4ee0-8453-03a3dfeeeb22