Implement overlays.
With this feature you can have a line style that is formed by over laying two lines.
By using a TYP file with appropriate definitions, you can create some useful effects.

To use this create an overlays file in the style.  Lines in the file look like this:

0x142: 0x12, 0x17, 0x18

This means that if you have a line of type 0x142, it will be written as three lines with
types 0x12, 0x17 and 0x18 respectively instead of 0x142.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@801 25d90789-57f7-4ee0-8453-03a3dfeeeb22