Add --transparent and --draw-priority options.
The --transparent option should make the map transparent, so that you can have two maps covering
the same area and see both of them.  Useful for contour lines among other things.

The --draw-priority is a number between 0 and 31 and it affects maps that cover the same area.
Maps with higher priorities are on top of maps with lower priorities.  I believe that transparent maps
are automatically on top of all non-transparent maps, but I am not an expert on this.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@836 25d90789-57f7-4ee0-8453-03a3dfeeeb22