- Code cleanup - Add option --pois-addresses to switch on the filling of POI address information
By default the address information is no longer added to the pois. Also
the filling is very conservative now and is not doing to much guessing. I
think in most countries this default is useful to minimize the risk of
wrong information. All POIs that have incomplete address info are marked
with "Fix me address" as street name. In countries with good is_in info
coverage I would suggest to use this options:

--pois-addresses --location-autofill=1

-Bernhard Heibler


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@871 25d90789-57f7-4ee0-8453-03a3dfeeeb22