Only let mapper know about last bounds element, not all.
Before, if the OSM file contained multiple bounds elements, the mapper
was told about each one and so the resulting map background rectangle
contained all of the areas. This meant that although the map features
were clipped to be within the area defined by the last bounds element,
the background rectangle was larger.

Now, when the OSM file contains multiple bounds elements, the background
rectangle will match the clipped features.

Obviously, this doesn't solve the general problem of coping with
multiple bounds elements.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@901 25d90789-57f7-4ee0-8453-03a3dfeeeb22