Make the sorting of cities repeatable. Previously I was getting a different order for the same input file each time I ran it.
We need to investigate how names should be sorted as it depends on locale
and not on character set.

For example ä comes at the end of the alphabet in Norwegian but is sorted
with a in some other languages.

This code sorts cities in the same way as cgpsmapper produced maps of 
germany I've seen.  The accented characters are sorted with the 
non-accented.  Previously accented characters were at the end.


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@959 25d90789-57f7-4ee0-8453-03a3dfeeeb22