Avoid setting access denied bits whose purpose we do not know when access=no.
The loop that set all the deny access bits was setting the two "unknown"
bits (delivery and emergency). As we don't really understand
what they do at this time, it's probably better not to set them.

Also add "taxi" to the list of vehicle classes so that the "deny taxi"
bit will get set when access=no.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@963 25d90789-57f7-4ee0-8453-03a3dfeeeb22