Add the 'continue' attribute to the type.
If the type has the continue attribute then searching is continued for further matches.
As a result the one feature in OSM can result in more than one element in the map.
These elements will be on top of each other (in the order found) and so can be used
for stylistic effect along with a suitable type file.

Also useful for ways that are tagged as two different kinds of things although this is
not the perfect solution to that problem.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1234 25d90789-57f7-4ee0-8453-03a3dfeeeb22