Impelent mdr 4 and 9 as best as we know.
The unknown byte in mdr4 is set to a constant, we still don't know what
it is there for.  Plenty of things appear to work still.

mdr10 is grouped based on the primary type and this grouping is indexed by mdr9.

The categories menu is now useful and will restrict the things returned.  Note that it is not
perfect yet, no matter what category you select you always get a couple of pubs and a bus
stop included in the results.  This is probably down to type/subtype confusion which
I will try to fix next.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1246 25d90789-57f7-4ee0-8453-03a3dfeeeb22