Add a character set aware label decoder.
This is a fairly big change as the CharacterDecoder interface is changed to have
a method that returns a DecodedText rather than an EncodedText.  The new class
DecodedText is more useful as it contains the actual string that you need.

non-ascii characters will now be correctly represented in the index, although they will still not
work as the sorting needs adjusting.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1266 25d90789-57f7-4ee0-8453-03a3dfeeeb22