Fill in the extra header values separately and controlled by the section class itself.
These values appear to control what fields appear in the section records and
their sizes.

I have confirmed by contstructing maps that the bottom two bits of the MDR 5 extra
value are the size of the local city field in mdr5 minus one, ie they are zero for a size of one.
This worked for all values of the bottom two bits.  The next two bits appear to relate the3byte label field, but
here only three (for the full size of the field) and zero for its complete absence appear to
work.

Anyway small maps now work.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1275 25d90789-57f7-4ee0-8453-03a3dfeeeb22