Reduce volume of warnings when turn restriction uses a 'via' way.
We don't yet (if ever) support via ways but when they occur in the map
data they cause inappropriate warning messages to be output. So even
if we can't actually generate a turn restriction with a via way, we can
at least cut down on the noise when they occur.

Patch from Marko Mäkelä.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1330 25d90789-57f7-4ee0-8453-03a3dfeeeb22