Ensure that cities and cityList have the same number of entries.
It would be nice to squash these two collections together sometime.

Also, if more than one city in a region/country has the same name then
roads or POIs that are near such cities are likely to end up referencing
the wrong city because the cities are looked up by name (that's a bug)!

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1390 25d90789-57f7-4ee0-8453-03a3dfeeeb22