Add support for disabling routing on unpaved roads and ferry routes.
By default, ways are considered to be paved and not ferry routes

Adding mkgmap:unpaved=yes/true/1 to a way will declare it to be unpaved
and such ways will not be routed through if the GPS has been set to avoid
unpaved roads.

Similarly, tagging a way with mkgmap:ferry=yes/true/1 will cause it to
be ignored for routing purposes if the GPS has been set to avoid ferry
routes.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1422 25d90789-57f7-4ee0-8453-03a3dfeeeb22