Fix regression introduced by r1435.
When scanning backwards looking for a place to split the way to limit
the way's number of points, don't set the flag that says we're splitting
the way at a previous node if the split is going to be at the current
point (i.e. we didn't need to go backwards at all).

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1436 25d90789-57f7-4ee0-8453-03a3dfeeeb22