Provide an alternative to using a multipolygon when generating the sea.
The --generate-sea option can now take a comma separated list of values:

multipolygon - generate the sea using a multipolygon (the default
behaviour so this really doesn't need to be specified).

polygons | no-mp - don't generate the sea using a multipolygon - instead,
generate a background sea polygon plus individual land polygons with
tag natural=land. This requires a suitable land polygon type to be defined
in the style file (suggested type is 0x010100) and the polygon must be
defined in the TYP file as having a higher drawing level than the sea
polygon type.

no-sea-sectors - disable the generation of "sea sectors" when the
coastline fails to reach the tile's boundary.

An unrecognised option value or the value "help" will cause some help
blurb to be printed.

For backwards compatibility, with no values specified, the --generate-sea
option should behave as it did before.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1449 25d90789-57f7-4ee0-8453-03a3dfeeeb22