Can now close small gaps in the coastline.
Added generate-sea sub-option, close-gaps, to specify the maximum length
(in metres) of gap that can be closed. Default behaviour is not to close
any gaps.

Gaps between coastline segments are closed with a straight line.

A separate line tagged natural=mkgmap:coastline-gap is also generated
so that you can see where the gaps have been closed.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1503 25d90789-57f7-4ee0-8453-03a3dfeeeb22