Extend sea generation code to cope with "anti-islands".
Anti-islands are closed polygons tagged with natural=coastline whose
direction is such that the water is on the inside and the land on the
outside. These are re-tagged natural=water so that they will be visible
above the land background.

If the tile contains at least one anti-island and no other coastline
reaches the edge of the tile, it is assumed that the tile background
should be land and not sea.

Anti-islands that are surrounded by water don't make sense so they are
converted into islands (they were probably meant to be islands but the
points were mistakenly drawn in reverse order).

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1527 25d90789-57f7-4ee0-8453-03a3dfeeeb22