The current speed optimisations from the style-speed branch.
There are no bug fixes, so any change in the maps produced is likely
to be a bug.

An index is created into the rules based on the first term in the
match expression.  A list of rules that might be matched is
constructed based on the tags present on each element and the possible
rules are looked up.

As rules can themselves add and modify tags, this is moderately
complex to get right.


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1562 25d90789-57f7-4ee0-8453-03a3dfeeeb22