Avoid flooding inland tiles due to bogus coastline segments.
When using --generate-sea=polygons, bad coastline segments
(e.g. anti-lakes) can cause the entire tile to be flooded with sea.

This commit suppresses the generation of the sea polygon if the coastline
doesn't reach the edge of the tile but ways with "highway" tags
(aka roads) do reach the edge of the tile.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1585 25d90789-57f7-4ee0-8453-03a3dfeeeb22