RuleSet: Update usedTags after merging other RuleSet
Issue: If you have a style A that has a base style B, only                      
the main tags from A are considered. This means, that many                      
rules from B are not used.                                                      

I haven't tracked down exactly why and where it happens,                        
but I assume, that some code cleans up the osm data                             
straight after/while reading it and removes any "useless"                       
tags. To save memory.                                                           

Anyway, I found a solution: Update the "usedTags" of the                        
RuleSet after merging in the base-RuleSet.                                      


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1654 25d90789-57f7-4ee0-8453-03a3dfeeeb22