Patch by WanMill that implements the proposal by Torsten Leitstikow :
My understanding of the multipolygons is, that the tags may EITHER be in the
relation OR on the outer polygons. So the outerpoylgons are only to be used,
when there are no tags on the relation.
If the relation itself is tagged and there is a tag on the outerpolygons, this
does logical mean, that the outerpolygon tags apply to the complete area
including the inner-area.

There shouldn't be a list of concerned tags, since any area tags may be used.
 
 - WanMil

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1711 25d90789-57f7-4ee0-8453-03a3dfeeeb22