Ensure that groups of mdr20 values are correct.
All cities with the same name should have the same mdr20 value. We step through and
ensure that this is so. The correct value is the lowest non-zero value that occurs in the group.

It would be good if we didn't need this extra complete transversal of the cities, but I don't think
it can be avoided without major changes.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1852 25d90789-57f7-4ee0-8453-03a3dfeeeb22