mdr5 and 20 changes
Set the mdr5 flag only when name and mdr20 value are the same.

Separate city records for cities with the same name but different regions as
well as different maps.

Now save the mdr20 value directly in the mdr5record. I don't think we need
to do any post processing of the values any more.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@1983 25d90789-57f7-4ee0-8453-03a3dfeeeb22