Create and use "lies in" information in preprocessed bounds
The "lies in" information of a preprocessed boundary defines in which other boundaries a given boundary is completely contained.
Example:
Boundary A10 (admin_level=10) lies in A8 (admin_level=8), A4 and A2. 
All elements in A10 are automatically in A8, A4 and A2. So the admin_levels 8, 4 and 2 are defined just be assigning the admin_level=10 boundary.

This has two advantages:
1. Perfomance: Time for index creation with bounds option can be reduced by 25% max.
2. Bugfix: Boundary information was inconsistent for elements located in two boundaries of the same level, e.g. a street crossing a city border and a region border could have created one city assigned to both regions. Now a city assigned by the bounds locator should always have the same region and country. 

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2013 25d90789-57f7-4ee0-8453-03a3dfeeeb22