Do not throw away polygons that potentially cover the complete bounding box.
The UnusedElementsRemoverHook now checks if the bounding box of a polygon covers the bbox of the tile completely. In this case it is possible that all points of the polygon lie outside the tile bbox but the polygon covers the tile completely. Especially this is true for the sea polygon when using --generate-sea=polygon. This was erroneously removed and therefore sea generation was broken using the polygon variant.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2209 25d90789-57f7-4ee0-8453-03a3dfeeeb22