Clear up all known bugs with the number preparer found by randomly generating number sequences.
It now basically completely works.

There is one problem, on certain streets that are composed of more than one segment,
the numbers will be present in all segments as shown by hovering the cursor along the road,
but the find function will not find numbers in one of the segments, instead taking you to
the nearest end of one of the working segments.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2407 25d90789-57f7-4ee0-8453-03a3dfeeeb22