Improve matching of street names for housenumber assignment
The housenumber handling uses the style sheet generated name of streets and matches that to the addr:street tag of housenumbers. The style sheet name is now stripped by the shield information (usually a ref) and text within brackets (also usually the ref name). This increases the match rate of housenumbers especially for streets with a ref.


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2654 25d90789-57f7-4ee0-8453-03a3dfeeeb22