Use mkgmap:street tag to get street name for house number search
The house number matching algorithm now checks the mkgmap:street tag on roads. If it is set it uses it to match to the addr:street tags of house number elements. This enables style developers to put refs and other infos into the street name and at the same time have a good matching for house number search.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2658 25d90789-57f7-4ee0-8453-03a3dfeeeb22