PartFilter extract 1) from end, 2) initial and final ranges.
Use negative numbers to take a part from the tail of the value list.  Use
the "<" or ">" operator (in addition to ":") in order to return a
concatenation of all the parts before or after a certain part. The
default behavior is not changed and this come at the expense of just 2
logical "IF" more when no operators are specified.

Example, if the value is "Aa#Bb#Cc#Dd#Ee"
part:#:1  returns Aa
part:#:-1 returns Ee
part:#:-2 returns Dd
part:#>1  returns Bb#Cc#Dd#Ee#
part:#<5  returns Aa#Bb#Cc#Dd#
part:#<-1 returns Aa#Bb#Cc#Dd#

- Enrico Liboni

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2668 25d90789-57f7-4ee0-8453-03a3dfeeeb22