Fix for 'mdr20 value changed' message.
Space was being treated as a non-character so any names that differed only
by a space would be treated the same during sorting although not during separation
into different streets while building the index.

This would be very unlikely to cause problem in normal circumstances, but when you have
another country with names that cannot be represented in 1250 then all the names turn into
sequences of question marks.  In that case is likely eventually that you will get something
like '??? ??' and '?? ???' followed by '??? ??' again, which sort the same but are seen as
different separating the streets while building the index.  The second '??? ??' causes a
problem because it has already been dealt with.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2704 25d90789-57f7-4ee0-8453-03a3dfeeeb22