Include style files relative to the including style.
The bug was that ff you include a file from the default style from within your own
style, then any files that are included within that file will be opened
relative to your style and not the default style.

Now included files are opened in the context of the style from which they are
included.



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2705 25d90789-57f7-4ee0-8453-03a3dfeeeb22