Avoid precalculation of area size for sea polygons created by the SeaGenerator.
For all sea polygons the area size was added as tag which increases the file size of the sea tiles but does not really help. The size only calculated the size of the sea in one sea tile and it's a bit random where the borders of a sea tile are. So the area size is not usefull.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@2971 25d90789-57f7-4ee0-8453-03a3dfeeeb22