optimization: reduce img size for shapes
We don't store the closing point of a shape. That means we don't
store the line that closes a shape. It saves a few Kb if 
we rotate the shape first so that it is the longest line
that is not saved, as this can have an influence on the number
of bits that is needed to store each point of the shape.
There is probably more potential for coastline shapes as
they often formed like a triangle with two long straight
lines and one curvy line. Adding a few points to the shorter
of the two straight lines could further reduce the needed bits
for one shape.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3008 25d90789-57f7-4ee0-8453-03a3dfeeeb22