apply rc-and-logging-v2.patch : set return code 1 if map creation failed
It also changes the handling of so called MapFailedExceptions like
"There is not enough room in a single garmin map..."
or
"Overflow of the NET1. The tile ..."

These messages are now printed as severe messages via the logging system (showing the input file):
Severe (MapFailedException): \ld_sp\63240003.o5m: (thrown in BufferedImgFileWriter.ensureSize()) There is not enough room in a single garmin map for all the input data. The .osm file should be split into smaller pieces first.

The stdout contains two new messages:
Number of MapFailedExceptions: 3
Number of ExitExceptions: 0

If any of the two shows a non-zero value, the program exits with return code 1.




git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3264 25d90789-57f7-4ee0-8453-03a3dfeeeb22