Keep multi byte character sets are char arrays.
So for sorts with the 'multi' flag set we take the char array directly
from the string to be sorted. For single byte code-pages we convert to
the single byte form and then create a char array with those values (so
that all the chars have the top byte equal to zero). May migrate everything
to use the unicode char values directly later, but this may create differences
with characters that are invalid for the code page.


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3279 25d90789-57f7-4ee0-8453-03a3dfeeeb22