Rename two argument remove() methods in MultiHashMap and MultiIdentityHashMap for java 8.
In java 8 there is a method of the same name but different signature on Map. So these
classes don't compile on java 8.  Renamed as removeMapping(), which is the same name
used by apache collections which faced a similar problem.

Note that the resulting map tiles are different when using java 7 and java 8. At least some of
this appears to be due to different iteration order in maps, but there may be more serious
problems - I don't know.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3314 25d90789-57f7-4ee0-8453-03a3dfeeeb22