Prevent a possible NullPointerException is house number generator
I could not reproduce the exception reported on the german OSM forum. But the check is required in any case. If an empty way is used as HousenumberMatch the getLocation() returns null which is not catched. Now use only elements that have getLocation() !=  null. Issue an error message in such a case.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3335 25d90789-57f7-4ee0-8453-03a3dfeeeb22