Improve ConvertFilter.
The patch allows for length conversions between m, km, mi, ft and
speed conversions of km/h (kmh, kmph), mph, knots. New ones can easily
be added.

It will also take into account any unit that already exists.

So for example if you have ${val|m=>ft}

Then val will be assumed to be in meters if there is no unit.

  val=10   --> 33

if a unit is given, then it converts from that, so if already in feet
the value will be unchanged.

  val=10ft --> 10
  val=10km --> 32,808

If it can't work out what to do then the value is unchanged. 

http://www.mkgmap.org.uk/pipermail/mkgmap-dev/2014q4/021995.html


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3353 25d90789-57f7-4ee0-8453-03a3dfeeeb22