Merge the mixed-index branch.
Each word of a street name is added to the global index.

To enable the functionality you need to use the --x-split-name-index
flag, since there is no logic to prevent useless parts of the name
(such as 'street', 'rue', 'calle' etc) being included. In the future
rules will be added so that it works by default according to the
country.



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3449 25d90789-57f7-4ee0-8453-03a3dfeeeb22