admin_centre-v1.patch: avoid to create duplicate POI
It implements the following:
When POIGeneratorHook creates a POI for a type=boundary relation
with boundary=administrative it searches for a role=admin_centre member
in that relation.
If one is found, the generated POI will use the coordinates of this member.

I see no easy way to compare the tags of the existing
node with those of the generated POI, so
as a second step, StyledConverter detects when a POI
with the same type and name (or empty name) is created at the
same Garmin coordinates (after style processing)
If that is true, the latter one is ignored and an info message is logged.

Maybe for certain types this should be changed to check for a radius
rather than equality, but that would require complex configuration.


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3573 25d90789-57f7-4ee0-8453-03a3dfeeeb22