dest_hint_v1.patch : change --process-destination option The process_destination option was a bit problematic because it could have added the tag destination=* to an existing OSM element.
The attached patch changes the method like this:
1) the tag destination is not changed by mkgmap
2) Instead the special tag mkgmap:dest_hint is now set to the
destination string that was found in one of the tags listed in this code snippet:
        tags.add("destination");
        tags.add("destination:lanes");
        tags.add("destination:lanes:forward");
        tags.add("destination:lanes:backward");
        tags.add("destination:forward");
        tags.add("destination:backward");
        tags.add("destination:street"); 


(BTW: This is also the order of evaluation in mkgmap searches since r3673, of cause
forward/backward are checked depending on the direction of the way)

For style authors this means that they have to
1) change all rules with mkgmap:dest_hint=true to mkgmap:dest_hint=*
2) change the rule that produces the hint to something like this:
mkgmap:dest_hint=*
  { set dest_hint = '${destination:ref|subst: =>} ${mkgmap:dest_hint|subst:;=> |subst:/=> }' |
         '${ref|subst: =>} ${mkgmap:dest_hint|subst:;=> |subst:/=> }' |
         '${mkgmap:dest_hint|subst:;=> |subst:/=> }';
       } 

Basically all places where "destination" was used were changed to mkgmap:dest_hint.
This could cause trouble, so I've added a check that complains when
the style contains an expression mkgmap:dest_hint=true.  

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3675 25d90789-57f7-4ee0-8453-03a3dfeeeb22