correct error that caused "internal error: via way is updated again with different nodes" error messages and ignored restriction relations
This message was printed when the style adds multiple routable ways for a via way of a restriction relation and when this via way is in the "wrong" direction (first point of via way is connected to the "to"-way, not the "from"-way.
This message might still be printed if the style adds such a via way with conflicting oneway attributes so that the order of points changes.

This also fixes a typo in method name isValidWithoputWay() -> isValidWithoutWay()

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3768 25d90789-57f7-4ee0-8453-03a3dfeeeb22