Correct style evaluation index
In special cases matching rules were not evaluated, e.g. for the following three rules
boundary=administrative { set mkgmap:if:0001=1 }
admin_level<3 & mkgmap:if:0001=1  [0x1e resolution 12]
mkgmap:if:0001!=1 & admin_level<3 [0x1f resolution 14]
a way with the tags boundary=xyz and admin_level=2 should trigger the last rule,
but it did not.

The problem was in the evaluation of so-called changable tags (mkgmap:if:0001 in this case). 
The effect of this correction is that the index returns more rules for evaluation
and probably a small increase in the run time.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3822 25d90789-57f7-4ee0-8453-03a3dfeeeb22