- add support for short syntax with two type definitions - make sure that rules inside an if / else block have at least a dummy expression like () so that we know where a new rule starts
A=1 & B=1 [0x1 resolution 24][0x10801 resolution 18]
# short form for
# A=1 & B=1 [0x1 resolution 24 continue]
# A=1 & B=1 [0x10801 resolution 18]

This no longer works:
if (expression) then
  {name ...} 
  [0x1 resolution]
end
as it was not clear if this means one ot two rules
You can use
if (expreession) then
  () {name ...} 
  [0x1 resolution]
end
if one rule was meant or 
if (expression) then
  () {name ...} 
  () [0x1 resolution]
end
for two rules. Of course this makes only sense if expression is a very complex term.

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3850 25d90789-57f7-4ee0-8453-03a3dfeeeb22