Add a rules stress test program
Generates rules at random and uses the style tester
to check if they a) compile and b) work correctly.

The baseline using current trunk is that with
10,000 rules, we get 205 failures, with 35 of
those being non-syntax errors.

By non-syntax I mean that the rule is accepted but
it either gives the wrong results, or it crashes
altogether.

The simplest rule that crashes mkgmap is:

  a>=1 & (a=1 | b~2) [0x2]



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@3984 25d90789-57f7-4ee0-8453-03a3dfeeeb22