implement option --x-dem-dists to allow multiple DEM zoom levels
sample usage for a map with 4 levels:
--x-dem-dists="5520, 16560, 44176, 88368" 
The steps in the values should somehow correspond to the steps in the levels option, but details are not fully understood yet.

The values are distances in DEM units. One DEM unit is 360° / 2^32.
The higher the value the smaller the DEM file.
HGT files contain either 1'' res which matches 3314 (2^32 / (360*3600)
or 3'' which matches 9942 (2^32 / (360 * 1200)

Without this option mkgmap calculates a single zoomLevel based on the best
resolution found in the hgt files.

Still missing: DEM for overview map

git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4012 25d90789-57f7-4ee0-8453-03a3dfeeeb22