implement experimental option --x-dem-set-extra to force crashes in MapSource
Only use for tests!!!

If given, mkgmap will always write the "extra" field.
If not, the exta field is only written when mkgmap has to encode
the "unknown" height value. The "unknown" height value is normally 
only encoded when the hgt file has voids (0x8000) or when
--dem-poly is used to set values outside of the polygon to this "unknown" value.
The extra flag seems to change the method that MapSource uses to 
extra the height values for the elevation graph ("Show Profile..." button). When the flag is set the DEM data must meet
more restrictions and the only good use of this option is to find those restrictions.



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4111 25d90789-57f7-4ee0-8453-03a3dfeeeb22