add code to detect and - maybe - remove overlapping way segments of routable lines.
A segment is a part of an OSM way build by pair of nodes.
The sequence of nodes only matters with oneways. 
Done:
- find overlapping segments in routable lines where two or more different OSM ways share the same segment
- split those ways into segments
- find overlapping segments with similar attributes (based on code in RoadMerger) and remove one if possible
- recombine the segments 
- remove those routable lines which are no longer complete and
replace them by copies with the newly combined segmnts

This happens after the treatment of unconnected roads and before the WrongAngleFixer starts.

TODO: 
- add code to handle route restrictions
- add code to handle ways with different labels
- add code to allow more differences, e.g. when
one way allows all vehicles and the other doesn't allow
bikes.



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4145 25d90789-57f7-4ee0-8453-03a3dfeeeb22