simplify and improve UnusedElementsRemoverHook (improved unused.patch published 2018-07-17)
1) The old code was not adapted to some changes in r3175:
It checks if a way without tags is member of a relation, because the rules in relation may add tags to this way.
This was okay when the hook was added to the source, but the hook is now executed after
those rules were executed, so that check is not needed.
2) It doesn't remove nodes outside of the tile bbox when such a node has tags keys that are used by the POI generator (--pois-to-areas-placement)
I've changed the code so that it checks for the combination of tag key and value, not only for the key.
3) The check if a way is outside of the tile bbox did not work for ways which have a bbox that intersects with the tile bbox
while the way itself neither crosses the bbox nor contains it (e.g. a large polygon).


git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4208 25d90789-57f7-4ee0-8453-03a3dfeeeb22