revert most changes from r4223 in WrongAngleFixer
This is a work around for a possible Assertion Error which can happen
when a node on a country boundary is very close to the tile boundary. The soft clipping process creates an additional node exactly on the tile boundary. This might cause a zig-zag.
The node on the tile boundary must not be replaced by the other one because the real clipping would add it again.



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4237 25d90789-57f7-4ee0-8453-03a3dfeeeb22