Improve typ file mapnik.txt
mapnik.patch by Ticker Berkin:

Patch attached that makes the following changes to mapnik.txt. This is
held in resources/typ-files and distributed in examples/typ-files.

- Add UTF-8 BOM at the start of the file, so that many text editors and
mkgmap will recognise the encoding.

- Add following line near start of file, for the same reasons:
; -*- coding: UTF-8 -*-

- comment out lines in the [_id] section; these are supplied /
overridden by mkgmap -- options

- Change polygon 0x3d / "Bay" to be transparent, with a low _drawOrder

- Add a default language label string to every entry. This is used by
the Garmin device when there is no label for the language the device is
set to.

The basis on which I've set the default labels is roughly as follows:

- Where my old Garmin device, without TYP file and for an unlabeled
object, gave a specific and unique label, used that.

- For other cases, eg where the standard Garmin label is something like
Road, Line, Area or Unnamed, used wording relevant to the OSM tags that
are being mapped.

- For some of the many Garmin Rivers and Lakes codes that we use for
other water areas, again used wording related to the OSM tag.

- Where the English translation (language code 0x04) is close enough to
the default label, I've deleted the English. I've also deleted the
English where I think the label (probably in all languages) doesn't
reflect the tag for which it is being used, but the default label does.

I've avoided making any changes to the visuals this TYP file produces
(apart from Bay), or to any language other than English.



git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4431 25d90789-57f7-4ee0-8453-03a3dfeeeb22